+++ stylis.js
@@ -45,12 +45,18 @@
 		var prefix = '';
 		var namespace = '';
 		var type = selector.charCodeAt(0) || 0;
+		
+		var char;
+		var attr;
+		var animns;
+		var plugins;
+		var uses;
 
 		// [ attr selector
 		if (type === 91) {
 			// `[data-id=namespace]` -> ['data-id', 'namespace']
-			var attr = selector.substring(1, selector.length-1).split('=');
-			var char = (namespace = attr[1]).charCodeAt(0);
+			attr = selector.substring(1, selector.length-1).split('=');
+			char = (namespace = attr[1]).charCodeAt(0);
 
 			// [data-id="namespace"]/[data-id='namespace']
 			// --> "namespace"/'namspace' --> namespace
@@ -72,8 +78,6 @@
 		// reset type signature
 		type = 0;
 
-		var animns;
-
 		// animation and keyframe namespace
 		if (animations == void 0 || animations === true) {
 			animations = true;
@@ -86,11 +90,10 @@
 
 		// uses middleware
 		var use = middleware != null;
-		var plugins;
 
 		// middleware
 		if (use) {
-			var uses = (typeof middleware).charCodeAt(0);
+			uses = (typeof middleware).charCodeAt(0);
 
 			// o, object of middlewares
 			if (uses === 111) {
@@ -126,6 +129,9 @@
 		var temp;
 		var prev;
 		var indexOf;
+		var first;
+		var second;
+		var third;
 
 		// variables
 		var variables;
@@ -140,12 +146,14 @@
 		var blck = '';
 		var nest = '';
 		var flat = '';
+		var code = 0;
 
 		// context signatures		
 		var special = 0;
 		var close = 0;
 		var closed = 0;
 		var comment = 0;
+		var comments = 0;
 		var strings = 0;
 		var nested = 0;
 		var func = 0;
@@ -179,13 +187,13 @@
 
 		// parse + compile
 		while (caret < eof) {
-			var code = styles.charCodeAt(caret);
+			code = styles.charCodeAt(caret);
 
 			// {, }, ; characters, parse line by line
 			if (strings === 0 && func === 0 && (code === 123 || code === 125 || code === 59)) {
 				buff += styles.charAt(caret);
 
-				var first = buff.charCodeAt(0);
+				first = buff.charCodeAt(0);
 
 				// only trim when the first character is a space ` `
 				if (first === 32) {
@@ -193,8 +201,8 @@
 				}
 
 				// default to 0 instead of NaN if there is no second/third character
-				var second = buff.charCodeAt(1) || 0;
-				var third = buff.charCodeAt(2) || 0;
+				second = buff.charCodeAt(1) || 0;
+				third = buff.charCodeAt(2) || 0;
 
 				// middleware, selector/property context, }
 				if (use && code !== 125) {
@@ -214,7 +222,9 @@
 
 				// ignore comments
 				if (comment === 2) {
-					code === 125 && (comment = 0);
+					if (code === 125) {
+						comment = 0;
+					}
 					buff = ''; 
 				}
 				// @, special block
@@ -289,7 +299,12 @@
 									char = styles.charCodeAt(caret);
 
 									// {, }, nested blocks may have nested blocks
-									char === 123 ? closed++ : char === 125 && closed--;
+									if (char === 123) {
+										closed++;
+									}
+									else if (char === 125) {
+										closed--;
+									}
 
 									// break when the nested block has ended
 									if (closed === 0) {
@@ -337,7 +352,7 @@
 							indexOf = buff.indexOf('(');
 
 							// function mixins
-							if (indexOf > -1) {
+							if (indexOf !== -1) {
 								// mixin name
 								var name = buff.substring(0, indexOf);
 
@@ -412,7 +427,9 @@
 					colon = buff.indexOf(':');
 
 					// first match create variables store 
-					variables === void 0 && (variables = []);
+					if (variables === void 0) {
+						variables = [];
+					}
 
 					// push key value pair
 					variables[variables.length] = [buff.substring(0, colon), buff.substring(colon+1, buff.length - 1).trim()];
@@ -447,7 +464,7 @@
 								for (var k = 0, l = props.length; k < l; k++) {
 									var prop = props[k].trim();
 									var frst = prop.charCodeAt(0);
-									var third = prop.charCodeAt(2);
+									var thrd = prop.charCodeAt(2);
 									var len = prop.length;
 									var last = prop.charCodeAt(len - 1);
 
@@ -457,47 +474,47 @@
 										last !== 41 && len !== 0 &&
 
 										// infinite, i, f, e
-										!(frst === 105 && third === 102 && last === 101 && len === 8) &&
+										!(frst === 105 && thrd === 102 && last === 101 && len === 8) &&
 
 										// linear, l, n, r
-										!(frst === 108 && third === 110 && last === 114 && len === 6) &&
+										!(frst === 108 && thrd === 110 && last === 114 && len === 6) &&
 
 										// alternate/alternate-reverse, a, t, e
-										!(frst === 97 && third === 116 && last === 101 && (len === 9 || len === 17)) &&
+										!(frst === 97 && thrd === 116 && last === 101 && (len === 9 || len === 17)) &&
 
 										// normal, n, r, l
-										!(frst === 110 && third === 114 && last === 108 && len === 6) &&
+										!(frst === 110 && thrd === 114 && last === 108 && len === 6) &&
 
 										// backwards, b, c, s
-										!(frst === 98 && third === 99 && last === 115 && len === 9) &&
+										!(frst === 98 && thrd === 99 && last === 115 && len === 9) &&
 
 										// forwards, f, r, s
-										!(frst === 102 && third === 114 && last === 115 && len === 8) &&
+										!(frst === 102 && thrd === 114 && last === 115 && len === 8) &&
 
 										// both, b, t, h
-										!(frst === 98 && third === 116 && last === 104 && len === 4) &&
+										!(frst === 98 && thrd === 116 && last === 104 && len === 4) &&
 
 										// none, n, n, e
-										!(frst === 110 && third === 110 && last === 101 && len === 4)&&
+										!(frst === 110 && thrd === 110 && last === 101 && len === 4)&&
 
 										// running, r, n, g 
-										!(frst === 114 && third === 110 && last === 103 && len === 7) &&
+										!(frst === 114 && thrd === 110 && last === 103 && len === 7) &&
 
 										// paused, p, u, d
-										!(frst === 112 && third === 117 && last === 100 && len === 6) &&
+										!(frst === 112 && thrd === 117 && last === 100 && len === 6) &&
 
 										// reversed, r, v, d
-										!(frst === 114 && third === 118 && last === 100 && len === 8) &&
+										!(frst === 114 && thrd === 118 && last === 100 && len === 8) &&
 
 										// step-start/step-end, s, e, (t/d)
 										!(
-											frst === 115 && third === 101 && 
+											frst === 115 && thrd === 101 && 
 											((last === 116 && len === 10) || (last === 100 && len === 8)) 
 										) &&
 
 										// ease/ease-in/ease-out/ease-in-out, e, s, e
 										!(
-											frst === 101 && third === 115 &&
+											frst === 101 && thrd === 115 &&
 											(
 												(last === 101 && len === 4) ||
 												(len === 11 || len === 7 || len === 8) && prop.charCodeAt(4) === 45
@@ -538,7 +555,7 @@
 					// display: d, i, s
 					else if (first === 100 && second === 105 && third === 115) {
 						// flex/inline-flex
-						if ((indexOf = buff.indexOf('flex')) > -1) {
+						if ((indexOf = buff.indexOf('flex')) !== -1) {
 							// e, inline-flex
 							temp = buff.charCodeAt(indexOf-2) === 101 ? 'inline-' : '';
 
@@ -660,7 +677,12 @@
 									char = styles.charCodeAt(caret);
 
 									// {, }, nested blocks may have nested blocks
-									char === 123 ? closed++ : char === 125 && closed--;
+									if (char === 123) {
+										closed++;
+									}
+									else if (char === 125) {
+										closed--;
+									}
 
 									// break when the nested block has ended
 									if (closed === 0) {
@@ -724,7 +746,7 @@
 									}
 
 									// [, [title="a,b,..."]
-									if (firstChar === 91) {
+									if (firstChar === 91 && selector.charCodeAt(selector.length-1) !== 93) {
 										for (var k = j+1, l = length-j; k < l; k++) {
 											var broken = (selector += ',' + selectors[k]).trim();
 
@@ -953,7 +975,21 @@
 				if (code === 13 || code === 10) {
 					// ignore line and block comments
 					if (comment === 2) {
-						buff = '';
+						// * character, block comment
+						if (buff.charCodeAt(buff.length - 2) === 42) {
+							buff = buff.substring(0, buff.indexOf('/*')).trim();
+						}
+						else {
+							// / character, does not start with `/`
+							if (buff.charCodeAt(0) !== 47 && (indexOf = buff.indexOf('//')) !== -1) {
+								buff = buff.substring(0, indexOf).trim();
+							}
+							else {
+								buff = '';
+							}
+						}
+
+						comments = 0;
 						comment = 0;
 					}
 
@@ -977,18 +1013,30 @@
 						}
 						// ( character
 						case 40: {
-							strings === 0 && (func = 1);
+							if (strings === 0) {
+								func = 1;
+							}
 							break;
 						}
 						// ) character
 						case 41: {
-							strings === 0 && (func = 0);
+							if (strings === 0) {
+								func = 0;
+							}
 							break;
 						}
 						// / character
 						case 47: {
-							if (strings === 0 && func !== 1) {
-								code === 47 && comment < 2 && comment++;
+							if (strings === 0 && func !== 1 && comment < 2) {
+								// * character
+								if (comments === 0 || styles.charCodeAt(caret - 1) === 42) {
+									comment++;				
+								}
+
+								// * character, allow line comments in block comments
+								if (comments === 0 && styles.charCodeAt(caret + 1) === 42) {
+									comments = 1;
+								}
 							}
 							break;
 						}